chore: codereview comments

+++ src/common.c
@@ -129,15 +129,15 @@ int parse_keepalive_conn_info(const uint8_t *buf, int len, connection_info_t *in
   if (version != SRTLA_KEEPALIVE_EXT_VERSION) return 0;
   
   // Parse connection info (all big-endian)
-  info->conn_id = (buf[14] << 24) | (buf[15] << 16) | (buf[16] << 8) | buf[17];
-  info->window = (int32_t)((buf[18] << 24) | (buf[19] << 16) | (buf[20] << 8) | buf[21]);
-  info->in_flight = (int32_t)((buf[22] << 24) | (buf[23] << 16) | (buf[24] << 8) | buf[25]);
+  info->conn_id = ((uint32_t)buf[14] << 24) | ((uint32_t)buf[15] << 16) | ((uint32_t)buf[16] << 8) | buf[17];
+  info->window = (int32_t)(((uint32_t)buf[18] << 24) | ((uint32_t)buf[19] << 16) | ((uint32_t)buf[20] << 8) | buf[21]);
+  info->in_flight = (int32_t)(((uint32_t)buf[22] << 24) | ((uint32_t)buf[23] << 16) | ((uint32_t)buf[24] << 8) | buf[25]);
   info->rtt_us = ((uint64_t)buf[26] << 56) | ((uint64_t)buf[27] << 48) | 
                  ((uint64_t)buf[28] << 40) | ((uint64_t)buf[29] << 32) |
                  ((uint64_t)buf[30] << 24) | ((uint64_t)buf[31] << 16) |
                  ((uint64_t)buf[32] << 8)  | (uint64_t)buf[33];
-  info->nak_count = (buf[34] << 24) | (buf[35] << 16) | (buf[36] << 8) | buf[37];
-  info->bitrate_bytes_per_sec = (buf[38] << 24) | (buf[39] << 16) | (buf[40] << 8) | buf[41];
+  info->nak_count = ((uint32_t)buf[34] << 24) | ((uint32_t)buf[35] << 16) | ((uint32_t)buf[36] << 8) | buf[37];
+  info->bitrate_bytes_per_sec = ((uint32_t)buf[38] << 24) | ((uint32_t)buf[39] << 16) | ((uint32_t)buf[40] << 8) | buf[41];
   
   return 1;
 }
